+++ styles/globals.scss
@@ -156,7 +156,7 @@ li,
       p {
         color: #444;
         margin-top: 2rem;
-        font-style: italic;
+        font-weight: 300;
       }
 
       .btns {
